feat(toast): allow clickable items without closing
Some content inside toasts should remain clickable without closing the toast. (even if closeOnClick is true)
While this was already hardcoded for a tags if fails for possible other tags like accordion or details (Especially when used together with the new #2197 😉 )
The previously hardcoded selector is not a customizable classname setting.

Btw: input is not needed to be set here, because any kind of input inside a toast will already set the whole toast to not closing when clicked (because it is expected to be clicked into the input fields).
The new setting here is supposed for elements which can be clicked, but are not expected to.