fix(form): make sure validation object is cloned
Whenever one has more than 2 forms on the page and also sets defaults to the global jquery forms object, any changes to the fields/rules afterwards were adopted to the global validation object instead of the dedicated one from the related form affecting all forms.

This is testable on the docs page fomantic-ui.com/behaviors/form.html

The first form "Tell us about yourself" did not have rules for gender and password anymore, although they are only removed for the fomantic-ui.com/behaviors/form.html#adding-rules-programmatically example only.
Whe you play with the "Add additional validation" and "Remove additional validation" buttons, it on the second mentioned example, this affect also all the other forms (whose have gender and/or password fields)

When applying the fix on the docs, each validation is dedicated to each form as it should.

This ia basically caused by jquery not seeming to doing a clone copy of a prototype default when extend is called without a deepcopy. Using extend also with deepcopy option solved it in the end.