fix(dropdown): use central delimiter char instead of key
On a multiple dropdown, having allowAdditions:true, the press of the delimiter key (comma by default) was checked via the keycode. However the keycode does not take possible pressed shift/alt keys into account.
This results in triggering a multi dropdown entry even when pressing the key together with shift/alt, so it was not possible to declare for example the semicolon as a delimiter.

As we already have the delimiter string for concatenation/splitting of existing values, this PR now also uses that setting for the keypress.