build(scripts): reduce nightly publishing
I've made changes to the nightly script to include the latest commit sha in the build metadata of the version so we can compare them. This allows us to only push the latest nightly to npm if there is now commits to push. I have also made changes to the workflow so it only continues and does the publish if the version was updated.