feat(chore): move from npm to yarn
Moving the project to yarn instead of npm.

Why?
These days yarn is used more often as it is faster then npm at package resolution in turn reducing the time spent setting up the dev environment. It also resolves packages better and because of this it can reduce disk space used.