fix(chore): use double colons for pseudo-elements
It's a minor nit-pick but as a FE Framework I think it's good to do things "the right way", even if one colon is probably also fine.

As a rule, double colons (::) should be used instead of a single colon (:). This distinguishes pseudo-classes from pseudo-elements. However, since this distinction was not present in older versions of the W3C spec, most browsers support both syntaxes for the original pseudo-elements.

developer.mozilla.org/en-US/docs/Web/CSS/Pseudo-elements