fix(dropdown): Hide submenus reliably, even with slow mouseleave detection
Some browsers seem to trigger mouse leave events only when moving the
mouse cursor. I noticed that on mobile browsers, especially when using
the touch screen. This somehow causes the submenus to stay visible even when
the root menu is hidden (e.g. due to clicking/selecting a menu item).

Thus, this commit forces hiding of all nested submenus, when hiding a menu.