fix(modal): avoid propagation of escape to modal
Allow using escape to dismiss calendars, dropdown, checkbox or search popups.
Previously, the escape key events (keydown, keyup) bubbled through the whole component tree.
This closed the modal when the escape key was used to dismiss a popup.

Fixing the issue requires all components to use the same event (keydown or keyup).
Standardize on keydown. This way propagation can be stopped prior to the modal closing.

This approach allows different components to "eat" the keystroke, while leaving the event
bubble up when required.