feat(dropdown): optional stuck addition item
When using a search dropdown, having allowAdditions activated and hideAdditions disabled, every time the user types something in the input, it is prepended in the menu.
Perhaps, in some cases, the user may not be aware that he has an addition option in the dropdown to click on :

there is no value selected yet in the dropdown. The user clicks on it, scrolls down, begins to type something, but there is so much results shown in the menu that the first line with the addition message displayed isn't shown
there is a value selected in the dropdown. The user clicks on it, and the value is displayed within the menu, but it is like the 50th element on 100. The scrollbar is half way. The user begins to type something else, and it doesn't show the first line with the addition message
It is even more remarkable if fullTextSearch equals true, a fuzzy search could display a lof of results while typing an addition.
If the user isn't used to that option, he may miss it or don't fully understand this option.

To activate this in dropdown, one has to add `className: {addition: 'stuck addition'}`