fix(search): blur on escape and proper type usage
Hitting the escape key is supposed to blur the search field, but this never happens. It only hides a possible given result set.
This PR now implements it the way it is meant to be and blurs the field when there is NO resultset being displayed (otherwise it still hides it as before)
I also fixed the set.type function properly. A possible given "type" parameter made no sense, was not used and it only worked with the default category. So now somebody really can create custom templates and a given classname is added to the search component , not only category.