fix(popup): detailed rtl check
I added more detailed check for rtl. Currently only the module is checked. However, if body or the surrounding context would be defined as rtl then the module itself is also displayed as rtl, so this PR enhances the check (i already did the same inside the sidebar PR and modal PR)