fix(dropdown): remove placeholderstyle when same as select text
When selecting the option from the dropdown that matches the default text value (e.g.<div class="default text">2</div>), the color of the selected value in the dropdown stays grey instead of turning black. This also applies if the value is already set (e.g. <input value="2" />). This is confusing as it states that the value has not been selected although it is. When reopening the dropdown menu the selected option is in bold and the value retrieved from the dropdown is correct. The issue is with the color not changing. And it only happens if the the option's text matches the default text.

If selecting other option and then again the option matching the default text - color is black as expected.