fix(build): one single component does not build
When trying to only compile one single component, for example doing this in semantic.json

"components": ["popup"],
the build does not compile anything. That's because the used micromatch/braces lib expects at least 2 entries so {popup,segment} would work, but {popup} does not. That leads to gulp not finding any file match, thus not compiling anything.