fix(modal): remove active class on hide
Also remove the active class when a modal is hidden. This makes sure the modal doesnt keep an in between added active class if hide was called when the show transition has not already ended. The class is already removed at the start of the hiding process and it is still needed there because of CSS settings which will otherwise disturb the hide animation.
This PR fixes such a specific use case