fix(theme): variable inheritance without default imported twice
This PR should finally fix proper variable inheritance without scoping and without overriding/double importing the default files twice.
This is solved by simply pointing the theme value to a non existing path in case it's default, because the default path is always imported anyway.

Another workaround would have been to just replace 'default' for each component inside theme.config by an empty string, but as theme.config is a custom file, nobody would have updated it for the next release.
I wanted this to be backward compatible, so i replaced the variable on the fly.

However, by using any LESS function to declare a variable (i tried replace/if/when/e/%), the access breaks when trying to use such variable via @{theme} inside mixins (which are scoped). :(
Even trying to trick it by using %('%s',@theme) instead of the string path did not work and resulted into "malformed import statement instead"
That said ( and even if this could be considered a bug in LESS), to solve all those issues, i just replaced the mixin call by importing another overrides.less file instead, which now isn't scoped anymore, so the import using the theme-variable also works there as well

Just in case somebody has made own custom components i left the loadUIOverrides mixin, but replaced the content by a comment warning so it's not a breaking change and people know what to do.