feat(dropdown): optional keeping the searchterm
New option keepSearchTerm for multiple search selection dropdowns. This (almost) reflects the current SUI behavior (which we fixed already)

The list stays filtered
The search term is not removed on item selection (to make it clear why the list is still filtered on selection)
This PR also removes the completely unused glyphWidth setting. This is all done in CSS instead of JS (not even in current SUI) and we fixed the search width glyph calculation already by #145 which in turn closes Semantic-Org/Semantic-UI#6655