Fixes for positions of icon, text and upward sub-menu shadow
Remove left padding from dropdown icon. Instead add right padding
to the dropdown item text since the icon display next to the text
when sub menu opens to the right direction as default.

Specify different shadow and rounded border for upward menu.

Increase the amount of top position value of remove icon of dropdown
to shift down to get better alignment along with the text and dropdown icon.

In previous version, the remove icon placed a few pixels above from
the text and dropdown icon which causes the inappropriate alignment.

Removed vertical alignment from the text container which causes
incorrect alignment with the elements around when it contains
the content which takes more space vertically such as an image.

Instead, use `align-item` property on the text container to have
everything inside of it in a linear position.

Remove unnecessary overrides and left position from menu which
causes the sub-menu positioned incorrectly when it's inside
floating dropdown item.

In vertical menu, the icon should be displayed above the text, so
the lowest order number should be specified.