fix(dropdown): support selection floating or pointing
When pointing or floating was used

background color overlayed the border corners of first and last item if menu was not scrolled/overflowed
in combination with selection borders were missing
inverted upward pointing had wrong arrow direction
I also added a new unlimited variant which does not make the selection menu scrollable but shows all items at once instead. This is especially needed in combination with pointing. Otherwise the arrow wont be displayed as the arrow is part of the (overflowing) dom node.