fix(input): allow shrink input along with the container width
The input box inside the input container always grow even the
container width is wide enough which causes the input field
overlaps the other field next to it when the screen is smaller.

This behavior is now changed by setting the input box to shrink
when the container is narrower in smaller screen.