feat(step): introduce new variant circular step
The new variant circular step shows a series of steps with the
circular marker and connection line between them.

It supports the three layout styles: horizontal, vertical and
ordered.

It has the same states as box step: active, completed, disabled.
And it contains the variations: color, inverted, size.

Due to it's design nature, it doesn't have the variations such
as: stackble, unstackable, attached and evenly divided.