fix(dropdown): fix map parameter on htmlentities
Multiple preselected values containing characters needed to be converted into htmlentities where encoded wrong (from the second value) because the related method was called as a .map() callback. However, .map callbacks are called given the index of the array as a second parameter. Unfortunately the given method expected something different which resulted in always encoding the ampersand character which broke the preselection