fix(checkbox): disallow keynav on readonly radios
A readyonly radio checkbox should not change its state via keynav from another radio of the same group.
This PR covers the situation when the readonly radio is already unchecked.

When it is already checked, it doesn't make sense to set this as the_only_ option to readonly as a radio is part of a radio group which, by nature, only has 1 active selection. And if one does not want this already selected option to be changed, all the other options from the same radio group should be either set to readonly or disabled.