fix(dropdown): set placeholdertext regardless of values
In situations where values are given for a dropdown, but the list is empty (or no value is selected), and also allowAdditions:true is set, a given placeholder wont be shown. That's because internally the dropdownlist does not get cleared when allowAdditions is true which also does not (re-)set the placeholder.

By always trying to show a placeholder this situation gets fixed.