fix(deps): remove re2 and url-regex through rework-import
I published a new version of fomantic/rework-import (through gulp-concat-css) which gets rid of re2 and url-regex(-safe) dependency.
The only need was to check a possible prototcol inside the import string. It does not matter if the whole URL is valid, as rework-import only needs to know if the import is a local file or not. This logic was already the case in 2.0 of the original rework-import, so basically there never was a need to switch to a separate package just for that check.

By doing all of this, fomantic does not need re2 dependency anymore which itself needs an installed python instance which especially on windows is not available by default.