fix(button): stackable needs fluid for 100% width
Stackable buttons always had a 100% width on mobile. While this might work in most cases, it doesnt when multiple buttons groups are used in the same row like floated buttons.

This PR removes the forced 100% width on mobile. If one needs to still have them being 100% wide, the fluid class is needed (which will evenly divide such buttons on non-mobile as well but doesnt break the ui appearance). Therefore i tag this as breaking change