fix(label): resolve the overlap issue during image transition
When using the ribbon label for the image card that features a visibility transition effect for lazy loading, it was observed that the ribbon label was partially overlapped by the image during the transition. This occurred because the ribbon is positioned absolutely without z-index value, which results the sibling image element to overlap.

This PR sets z-index value to the ribbon label within the image card which resolve the overlap issue.