fix(modal, flyout, sidebar): remove style, if initial bodymargin was 0
When a modal/flyout/sidebar is opened, it adjusts the body (or context) margin to avoid moving the modal by the width of a possible scrollbar.
When the modal closes, it reverts the value back to the initial one.
However, the default of margin/padding if 0, so there is no need to keep the inline style to 0. Instead the style attribute can be simply removed. This solves some issues in cases people use custom classes which adjusts the body margin as well