fix(dropdown): prevent change loop trigger
Whenever a dropdown selectton should be cleared, we added a check for existing labels to remove them by #3000 in case the input value was modified manually outside of fui logic.

This worked as long as there isnt any additional event listener outside of fomantic which itself does again changes to the dropdown input itself, like the example from #3382 using vuejs to generate the option values and listen to their changes and modifies/retriggers the change event aga.

In such cases we now explicitely prevent the change trigger for additional set values after the removal of existing labels.
Rare case, but possible.

This PR also fixes a regression of #3344 where includes was sometimes performed on a boolean value which in turn results in a console error